requirements.txt
python-dotenv
pandas
fastapi
pydantic
langchain-google-genai
langchain
langchain-community
sentence-transformers
faiss-gpu
uvicorn